Update PAA lista1 and lista2

+++ disciplinas/PAA/listas/lista_2/exercicios/exercicio_1.cpp
@@ -11,7 +11,7 @@ using namespace std;
 
 int main()
 {
-    vector<int> arr = ReadingFile();
+    vector<float> arr = ReadingFile();
     PrintVector(arr);
     BubbleSort(arr);
     PrintVector(arr);
