+++ disciplinas/PAA/listas/lista_2/exercicios/exercicio_2.cpp
@@ -11,7 +11,7 @@ using namespace std;
 
 int main()
 {
-    vector<int> arr = ReadingFile();
+    vector<float> arr = ReadingFile();
     PrintVector(arr);
     SelectionSort(arr);
     PrintVector(arr);
